+++ resource/api/main.py
@@ -58,6 +58,7 @@ def consumption(
 ):
     if x_amzn_mtls_clientcert is None:
         raise HTTPException(status_code=401, detail="No client certificate provided")
+    print(x_amzn_mtls_clientcert)
     if not auth.require_role("carbon-accounting@perseus", x_amzn_mtls_clientcert):
         raise HTTPException(
             status_code=403,
